Starts trojan handling for CSR/trap vector handling
Trap vector delegation depends explicitly on whether the core is in M or
S mode. As the trojan mode E will destroy the ability for the processor
to correctly track this we add duplicate tracking specifically for parts
of CSR/trap vector handling which depend on M versus S modes

This commit is unfinished as this duplicated tracking still dependes on
the `STATUS_MPP`/`STATUS_SPP` CSRs, which will also have faulty values when
in E mode, and so we will also need to be duplicated.

We should also ensure that the `TrapM`, `mretM`, `sretM` wires are
correctly produced even when in E mode.